fixed 65535 pixels width/height limitation in FreeImage_Paste

+++ FreeImage/Source/FreeImageToolkit/CopyPaste.cpp
@@ -41,19 +41,19 @@
 
 // ----------------------------------------------------------
 /// 1-bit
-static BOOL Combine1(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha);
+static BOOL Combine1(FIBITMAP *dst_dib, FIBITMAP *src_dib, unsigned x, unsigned y, unsigned alpha);
 /// 4-bit
-static BOOL Combine4(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha);
+static BOOL Combine4(FIBITMAP *dst_dib, FIBITMAP *src_dib, unsigned x, unsigned y, unsigned alpha);
 /// 8-bit
-static BOOL Combine8(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha);
+static BOOL Combine8(FIBITMAP *dst_dib, FIBITMAP *src_dib, unsigned x, unsigned y, unsigned alpha);
 /// 16-bit 555
-static BOOL Combine16_555(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha);
+static BOOL Combine16_555(FIBITMAP *dst_dib, FIBITMAP *src_dib, unsigned x, unsigned y, unsigned alpha);
 /// 16-bit 565
-static BOOL Combine16_565(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha);
+static BOOL Combine16_565(FIBITMAP *dst_dib, FIBITMAP *src_dib, unsigned x, unsigned y, unsigned alpha);
 /// 24-bit
-static BOOL Combine24(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha);
+static BOOL Combine24(FIBITMAP *dst_dib, FIBITMAP *src_dib, unsigned x, unsigned y, unsigned alpha);
 /// 32- bit
-static BOOL Combine32(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha);
+static BOOL Combine32(FIBITMAP *dst_dib, FIBITMAP *src_dib, unsigned x, unsigned y, unsigned alpha);
 // ----------------------------------------------------------
 
 // ----------------------------------------------------------
@@ -61,7 +61,7 @@ static BOOL Combine32(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD
 // ----------------------------------------------------------
 
 static BOOL 
-Combine1(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
+Combine1(FIBITMAP *dst_dib, FIBITMAP *src_dib, unsigned x, unsigned y, unsigned alpha) {
 	BOOL value;
 
 	// check the bit depth of src and dst images
@@ -78,8 +78,8 @@ Combine1(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
 	BYTE *src_bits = FreeImage_GetBits(src_dib);	
 
 	// combine images
-	for(WORD rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
-		for(WORD cols = 0; cols < FreeImage_GetWidth(src_dib); cols++) {
+	for(unsigned rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
+		for(unsigned cols = 0; cols < FreeImage_GetWidth(src_dib); cols++) {
 			// get bit at (rows, cols) in src image
 			value = (src_bits[cols >> 3] & (0x80 >> (cols & 0x07))) != 0;
 			// set bit at (rows, x+cols) in dst image	
@@ -98,7 +98,7 @@ Combine1(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
 // ----------------------------------------------------------
 
 static BOOL 
-Combine4(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
+Combine4(FIBITMAP *dst_dib, FIBITMAP *src_dib, unsigned x, unsigned y, unsigned alpha) {
 
 	int swapTable[16];
 	BOOL bOddStart, bOddEnd;
@@ -167,12 +167,12 @@ Combine4(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
 		bOddEnd = FALSE;
 	}
 
-	for(WORD rows = 0; rows < src_height; rows++) {
+	for(unsigned rows = 0; rows < src_height; rows++) {
 		memcpy(buffer, src_bits, src_line);
 		
 		// change the values in the temp row to be those from the swap table
 		
-		for (WORD cols = 0; cols < src_line; cols++) {
+		for (unsigned cols = 0; cols < src_line; cols++) {
 			buffer[cols] = (BYTE)((swapTable[HINIBBLE(buffer[cols]) >> 4] << 4) + swapTable[LOWNIBBLE(buffer[cols])]);
 		}
 
@@ -201,7 +201,7 @@ Combine4(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
 // ----------------------------------------------------------
 
 static BOOL 
-Combine8(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
+Combine8(FIBITMAP *dst_dib, FIBITMAP *src_dib, unsigned x, unsigned y, unsigned alpha) {
 	// check the bit depth of src and dst images
 	if((FreeImage_GetBPP(dst_dib) != 8) || (FreeImage_GetBPP(src_dib) != 8)) {
 		return FALSE;
@@ -217,7 +217,7 @@ Combine8(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
 
 	if(alpha > 255) {
 		// combine images
-		for(WORD rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
+		for(unsigned rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
 			memcpy(dst_bits, src_bits, FreeImage_GetLine(src_dib));
 
 			dst_bits += FreeImage_GetPitch(dst_dib);
@@ -225,8 +225,8 @@ Combine8(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
 		}
 	} else {
 		// alpha blend images
-		for(WORD rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
-			for (WORD cols = 0; cols < FreeImage_GetLine(src_dib); cols++) {							
+		for(unsigned rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
+			for (unsigned cols = 0; cols < FreeImage_GetLine(src_dib); cols++) {							
 				dst_bits[cols] = (BYTE)(((src_bits[cols] - dst_bits[cols]) * alpha + (dst_bits[cols] << 8)) >> 8);
 			}
 
@@ -243,7 +243,7 @@ Combine8(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
 // ----------------------------------------------------------
 
 static BOOL 
-Combine16_555(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
+Combine16_555(FIBITMAP *dst_dib, FIBITMAP *src_dib, unsigned x, unsigned y, unsigned alpha) {
 	// check the bit depth of src and dst images
 	if((FreeImage_GetBPP(dst_dib) != 16) || (FreeImage_GetBPP(src_dib) != 16)) {
 		return FALSE;
@@ -258,15 +258,15 @@ Combine16_555(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) 
 	BYTE *src_bits = FreeImage_GetBits(src_dib);	
 
 	if (alpha > 255) {
-		for(WORD rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
+		for(unsigned rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
 			memcpy(dst_bits, src_bits, FreeImage_GetLine(src_dib));
 
 			dst_bits += FreeImage_GetPitch(dst_dib);
 			src_bits += FreeImage_GetPitch(src_dib);
 		}
 	} else {
-		for(WORD rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
-			for(WORD cols = 0; cols < FreeImage_GetLine(src_dib); cols += 2) {
+		for(unsigned rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
+			for(unsigned cols = 0; cols < FreeImage_GetLine(src_dib); cols += 2) {
 				RGBTRIPLE color_s;
 				RGBTRIPLE color_t;
 				
@@ -303,7 +303,7 @@ Combine16_555(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) 
 }
 
 static BOOL 
-Combine16_565(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
+Combine16_565(FIBITMAP *dst_dib, FIBITMAP *src_dib, unsigned x, unsigned y, unsigned alpha) {
 	// check the bit depth of src and dst images
 	if((FreeImage_GetBPP(dst_dib) != 16) || (FreeImage_GetBPP(src_dib) != 16)) {
 		return FALSE;
@@ -318,15 +318,15 @@ Combine16_565(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) 
 	BYTE *src_bits = FreeImage_GetBits(src_dib);	
 
 	if (alpha > 255) {
-		for(WORD rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
+		for(unsigned rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
 			memcpy(dst_bits, src_bits, FreeImage_GetLine(src_dib));
 
 			dst_bits += FreeImage_GetPitch(dst_dib);
 			src_bits += FreeImage_GetPitch(src_dib);
 		}
 	} else {
-		for(WORD rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
-			for(WORD cols = 0; cols < FreeImage_GetLine(src_dib); cols += 2) {
+		for(unsigned rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
+			for(unsigned cols = 0; cols < FreeImage_GetLine(src_dib); cols += 2) {
 				RGBTRIPLE color_s;
 				RGBTRIPLE color_t;
 				
@@ -367,7 +367,7 @@ Combine16_565(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) 
 // ----------------------------------------------------------
 
 static BOOL 
-Combine24(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
+Combine24(FIBITMAP *dst_dib, FIBITMAP *src_dib, unsigned x, unsigned y, unsigned alpha) {
 	// check the bit depth of src and dst images
 	if((FreeImage_GetBPP(dst_dib) != 24) || (FreeImage_GetBPP(src_dib) != 24)) {
 		return FALSE;
@@ -383,7 +383,7 @@ Combine24(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
 
 	if(alpha > 255) {
 		// combine images
-		for(WORD rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
+		for(unsigned rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
 			memcpy(dst_bits, src_bits, FreeImage_GetLine(src_dib));
 
 			dst_bits += FreeImage_GetPitch(dst_dib);
@@ -391,8 +391,8 @@ Combine24(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
 		}
 	} else {
 		// alpha blend images
-		for(WORD rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
-			for (WORD cols = 0; cols < FreeImage_GetLine(src_dib); cols++) {							
+		for(unsigned rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
+			for (unsigned cols = 0; cols < FreeImage_GetLine(src_dib); cols++) {							
 				dst_bits[cols] = (BYTE)(((src_bits[cols] - dst_bits[cols]) * alpha + (dst_bits[cols] << 8)) >> 8);
 			}
 
@@ -409,7 +409,7 @@ Combine24(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
 // ----------------------------------------------------------
 
 static BOOL 
-Combine32(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
+Combine32(FIBITMAP *dst_dib, FIBITMAP *src_dib, unsigned x, unsigned y, unsigned alpha) {
 	// check the bit depth of src and dst images
 	if((FreeImage_GetBPP(dst_dib) != 32) || (FreeImage_GetBPP(src_dib) != 32)) {
 		return FALSE;
@@ -425,7 +425,7 @@ Combine32(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
 
 	if (alpha > 255) {
 		// combine images
-		for(WORD rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
+		for(unsigned rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
 			memcpy(dst_bits, src_bits, FreeImage_GetLine(src_dib));
 
 			dst_bits += FreeImage_GetPitch(dst_dib);
@@ -433,8 +433,8 @@ Combine32(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
 		}
 	} else {
 		// alpha blend images
-		for(WORD rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
-			for(WORD cols = 0; cols < FreeImage_GetLine(src_dib); cols++) {
+		for(unsigned rows = 0; rows < FreeImage_GetHeight(src_dib); rows++) {
+			for(unsigned cols = 0; cols < FreeImage_GetLine(src_dib); cols++) {
 				dst_bits[cols] = (BYTE)(((src_bits[cols] - dst_bits[cols]) * alpha + (dst_bits[cols] << 8)) >> 8);
 			}
 
@@ -451,19 +451,19 @@ Combine32(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y, WORD alpha) {
 // ----------------------------------------------------------
 
 static BOOL 
-CombineSameType(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y) {
+CombineSameType(FIBITMAP *dst_dib, FIBITMAP *src_dib, unsigned x, unsigned y) {
 	// check the bit depth of src and dst images
 	if(FreeImage_GetImageType(dst_dib) != FreeImage_GetImageType(src_dib)) {
 		return FALSE;
 	}
 
-	int src_width  = FreeImage_GetWidth(src_dib);
-	int src_height = FreeImage_GetHeight(src_dib);
-	int src_pitch  = FreeImage_GetPitch(src_dib);
-	int src_line   = FreeImage_GetLine(src_dib);
-	int dst_width  = FreeImage_GetWidth(dst_dib);
-	int dst_height = FreeImage_GetHeight(dst_dib);
-	int dst_pitch  = FreeImage_GetPitch(dst_dib);
+	unsigned src_width  = FreeImage_GetWidth(src_dib);
+	unsigned src_height = FreeImage_GetHeight(src_dib);
+	unsigned src_pitch  = FreeImage_GetPitch(src_dib);
+	unsigned src_line   = FreeImage_GetLine(src_dib);
+	unsigned dst_width  = FreeImage_GetWidth(dst_dib);
+	unsigned dst_height = FreeImage_GetHeight(dst_dib);
+	unsigned dst_pitch  = FreeImage_GetPitch(dst_dib);
 	
 	// check the size of src image
 	if((x + src_width > dst_width) || (y + src_height > dst_height)) {
@@ -474,7 +474,7 @@ CombineSameType(FIBITMAP *dst_dib, FIBITMAP *src_dib, WORD x, WORD y) {
 	BYTE *src_bits = FreeImage_GetBits(src_dib);	
 
 	// combine images	
-	for(WORD rows = 0; rows < src_height; rows++) {
+	for(unsigned rows = 0; rows < src_height; rows++) {
 		memcpy(dst_bits, src_bits, src_line);
 
 		dst_bits += dst_pitch;
@@ -695,27 +695,27 @@ FreeImage_Paste(FIBITMAP *dst, FIBITMAP *src, int left, int top, int alpha) {
 		// paste src to dst
 		switch(FreeImage_GetBPP(dst)) {
 			case 1:
-				bResult = Combine1(dst, clone, (WORD)left, (WORD)top, (WORD)alpha);
+				bResult = Combine1(dst, clone, (unsigned)left, (unsigned)top, (unsigned)alpha);
 				break;
 			case 4:
-				bResult = Combine4(dst, clone, (WORD)left, (WORD)top, (WORD)alpha);
+				bResult = Combine4(dst, clone, (unsigned)left, (unsigned)top, (unsigned)alpha);
 				break;
 			case 8:
-				bResult = Combine8(dst, clone, (WORD)left, (WORD)top, (WORD)alpha);
+				bResult = Combine8(dst, clone, (unsigned)left, (unsigned)top, (unsigned)alpha);
 				break;
 			case 16:
 				if (isRGB565) {
-					bResult = Combine16_565(dst, clone, (WORD)left, (WORD)top, (WORD)alpha);
+					bResult = Combine16_565(dst, clone, (unsigned)left, (unsigned)top, (unsigned)alpha);
 				} else {
 					// includes case where all the masks are 0
-					bResult = Combine16_555(dst, clone, (WORD)left, (WORD)top, (WORD)alpha);
+					bResult = Combine16_555(dst, clone, (unsigned)left, (unsigned)top, (unsigned)alpha);
 				}
 				break;
 			case 24:
-				bResult = Combine24(dst, clone, (WORD)left, (WORD)top, (WORD)alpha);
+				bResult = Combine24(dst, clone, (unsigned)left, (unsigned)top, (unsigned)alpha);
 				break;
 			case 32:
-				bResult = Combine32(dst, clone, (WORD)left, (WORD)top, (WORD)alpha);
+				bResult = Combine32(dst, clone, (unsigned)left, (unsigned)top, (unsigned)alpha);
 				break;
 		}
 
@@ -724,7 +724,7 @@ FreeImage_Paste(FIBITMAP *dst, FIBITMAP *src, int left, int top, int alpha) {
 
 		}
 	else {	// any type other than FITBITMAP
-		bResult = CombineSameType(dst, src, (WORD)left, (WORD)top);
+		bResult = CombineSameType(dst, src, (unsigned)left, (unsigned)top);
 	}
 
 	return bResult;
